CORE-14685: Separate UniquenessChecker and BackingStore from Lifecycle. (#4251)
