+++ components/uniqueness/backing-store-impl/src/main/kotlin/net/corda/uniqueness/backingstore/impl/JPABackingStoreImpl.kt
@@ -6,17 +6,8 @@ import net.corda.db.connection.manager.DbConnectionManager
 import net.corda.db.connection.manager.VirtualNodeDbType
 import net.corda.db.core.DbPrivilege
 import net.corda.db.schema.CordaDb
-import net.corda.lifecycle.DependentComponents
-import net.corda.lifecycle.LifecycleCoordinator
-import net.corda.lifecycle.LifecycleCoordinatorFactory
-import net.corda.lifecycle.LifecycleEvent
-import net.corda.lifecycle.RegistrationStatusChangeEvent
-import net.corda.lifecycle.StartEvent
-import net.corda.lifecycle.StopEvent
-import net.corda.lifecycle.createCoordinator
 import net.corda.metrics.CordaMetrics
 import net.corda.orm.JpaEntitiesRegistry
-import net.corda.orm.JpaEntitiesSet
 import net.corda.uniqueness.backingstore.BackingStore
 import net.corda.uniqueness.datamodel.common.UniquenessConstants.HIBERNATE_JDBC_BATCH_SIZE
 import net.corda.uniqueness.datamodel.common.UniquenessConstants.RESULT_ACCEPTED_REPRESENTATION
@@ -28,7 +19,6 @@ import net.corda.uniqueness.datamodel.impl.UniquenessCheckStateDetailsImpl
 import net.corda.uniqueness.datamodel.impl.UniquenessCheckStateRefImpl
 import net.corda.uniqueness.datamodel.internal.UniquenessCheckRequestInternal
 import net.corda.uniqueness.datamodel.internal.UniquenessCheckTransactionDetailsInternal
-import net.corda.utilities.VisibleForTesting
 import net.corda.utilities.debug
 import net.corda.v5.application.uniqueness.model.UniquenessCheckError
 import net.corda.v5.application.uniqueness.model.UniquenessCheckResult
@@ -55,8 +45,6 @@ import javax.persistence.RollbackException
  */
 @Component(service = [BackingStore::class])
 open class JPABackingStoreImpl @Activate constructor(
-    @Reference(service = LifecycleCoordinatorFactory::class)
-    coordinatorFactory: LifecycleCoordinatorFactory,
     @Reference(service = JpaEntitiesRegistry::class)
     private val jpaEntitiesRegistry: JpaEntitiesRegistry,
     @Reference(service = DbConnectionManager::class)
@@ -70,17 +58,12 @@ open class JPABackingStoreImpl @Activate constructor(
         const val MAX_ATTEMPTS = 10
     }
 
-    private val lifecycleCoordinator: LifecycleCoordinator = coordinatorFactory
-        .createCoordinator<BackingStore>(::eventHandler)
-
-    private val dependentComponents = DependentComponents.of(
-        ::dbConnectionManager
-    )
-
-    private lateinit var jpaEntities: JpaEntitiesSet
-
-    override val isRunning: Boolean
-        get() = lifecycleCoordinator.isRunning
+    init {
+        jpaEntitiesRegistry.register(
+            CordaDb.Uniqueness.persistenceUnitName,
+            JPABackingStoreEntities.classes
+        )
+    }
 
     override fun session(holdingIdentity: HoldingIdentity, block: (BackingStore.Session) -> Unit) {
 
@@ -116,16 +99,6 @@ open class JPABackingStoreImpl @Activate constructor(
                 .build()
                 .record(Duration.ofNanos(System.nanoTime() - sessionStartTime))
         }
-    }
-
-    override fun start() {
-        log.info("Backing store starting")
-        lifecycleCoordinator.start()
-    }
-
-    override fun stop() {
-        log.info("Backing store stopping")
-        lifecycleCoordinator.stop()
     }
 
     protected open inner class SessionImpl(
@@ -430,37 +403,6 @@ open class JPABackingStoreImpl @Activate constructor(
                     .withTag(CordaMetrics.Tag.SourceVirtualNode, holdingIdentity.shortHash.toString())
                     .build()
                     .record(Duration.ofNanos(System.nanoTime() - commitStartTime))
-            }
-        }
-    }
-
-    @VisibleForTesting
-    fun eventHandler(event: LifecycleEvent, coordinator: LifecycleCoordinator) {
-        log.info("Backing store received event $event")
-        when (event) {
-            is StartEvent -> {
-                dependentComponents.registerAndStartAll(coordinator)
-            }
-            is StopEvent -> {
-                dependentComponents.stopAll()
-            }
-            is RegistrationStatusChangeEvent -> {
-                jpaEntitiesRegistry.register(
-                    CordaDb.Uniqueness.persistenceUnitName,
-                    JPABackingStoreEntities.classes
-                )
-
-                jpaEntities = jpaEntitiesRegistry.get(CordaDb.Uniqueness.persistenceUnitName)
-                    ?: throw IllegalStateException(
-                        "persistenceUnitName " +
-                                "${CordaDb.Uniqueness.persistenceUnitName} is not registered."
-                    )
-
-                log.info("Backing store is ${event.status}")
-                coordinator.updateStatus(event.status)
-            }
-            else -> {
-                log.warn("Unexpected event ${event}, ignoring")
             }
         }
     }
