+++ components/uniqueness/uniqueness-checker-client-service-impl/build.gradle
@@ -10,7 +10,7 @@ dependencies {
     compileOnly 'org.osgi:osgi.core'
     compileOnly 'org.osgi:org.osgi.service.component.annotations'
 
-    implementation platform('net.corda:corda-api:$cordaApiVersion')
+    implementation platform("net.corda:corda-api:$cordaApiVersion")
 
     implementation 'net.corda:corda-application'
     implementation 'net.corda:corda-base'
