+++ components/uniqueness/uniqueness-checker-client-service-impl/src/main/kotlin/net/corda/uniqueness/client/impl/UniquenessCheckClientExternalEventFactory.kt
@@ -10,12 +10,11 @@ import net.corda.schema.Schemas
 import net.corda.uniqueness.datamodel.common.toUniquenessResult
 import net.corda.v5.application.uniqueness.model.UniquenessCheckResult
 import net.corda.virtualnode.toAvro
-import org.osgi.service.component.annotations.Activate
 import org.osgi.service.component.annotations.Component
 import java.time.Instant
 
 @Component(service = [ExternalEventFactory::class])
-class UniquenessCheckExternalEventFactory @Activate constructor():
+class UniquenessCheckExternalEventFactory :
     ExternalEventFactory<UniquenessCheckExternalEventParams, UniquenessCheckResponseAvro, UniquenessCheckResult> {
 
     override val responseType = UniquenessCheckResponseAvro::class.java
