+++ components/uniqueness/uniqueness-checker-impl/build.gradle
@@ -17,10 +17,10 @@ dependencies {
     implementation 'net.corda:corda-topic-schema'
     implementation platform("net.corda:corda-api:$cordaApiVersion")
 
+    api project(":components:uniqueness:uniqueness-checker")
     implementation project(":components:configuration:configuration-read-service")
     implementation project(":components:uniqueness:backing-store")
     implementation project(":components:uniqueness:backing-store-impl")
-    implementation project(":components:uniqueness:uniqueness-checker")
     implementation project(':libs:configuration:configuration-core')
     implementation project(':libs:configuration:configuration-datamodel')
     implementation project(':libs:flows:external-event-responses')
